server/requirements.txt
livekit-agents
livekit-plugins-openai
python-dotenv
livekit-plugins-deepgram
livekit-plugins-silero
livekit-plugins-google
livekit-plugins-